Pull the banner‑display logic out into its own module (settings/banner.js) and wire both the Options UI and the Background script to import it.
